Remove getPrice from CatalogService as it was unused
